+++ modules/liquidstaking/source/local/source.go
@@ -6,7 +6,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/forbole/juno/v4/node/local"
 
-	liquidstakingtypes "github.com/crescent-network/crescent/v4/x/liquidstaking/types"
+	liquidstakingtypes "github.com/crescent-network/crescent/v5/x/liquidstaking/types"
 	liquidstakingsource "github.com/forbole/bdjuno/v4/modules/liquidstaking/source"
 )